--- Gestione/Controllers/EspLav.cs
@@ -1,31 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using System.Web.Mvc;
-using Gestione.Models;
-using Interfaces;
-
-namespace Gestione.Controllers {
-    public partial class HomeController : Controller {
-        [HttpPost]
-        public ActionResult ModEspLav(int annoInizioEsp, int annoFineEsp, string qualifica, string descrizioneEsp){
-            EspLav esp = new EspLav{ AnnoInizio=annoInizioEsp,AnnoFine=annoFineEsp,Qualifica=qualifica,Descrizione=descrizioneEsp};
-            Profilo p = Session["profile"] as Profilo;
-            EspLav espV = Session["esperienza"] as EspLav;
-            dm.ModEspLav(p.Matricola,espV,esp);
-            ViewBag.Message = "Funziona";
-            return View($"MyPage");
-        }
-
-        public ActionResult AddEspLav(int annoinizioesp, int annofinesp, string qualifica, string descrizionesp){
-            EspLav esp = new EspLav{ AnnoInizio=annoinizioesp,AnnoFine= annofinesp,Qualifica=qualifica,Descrizione=descrizionesp};
-            Profilo p = Session["profile"] as Profilo;
-            dm.AddEspLav(p.Matricola,esp);
-            ViewBag.Message="Esperienza aggiunta nel curriculum,corri a controllare!";
-            return View($"MyPage");
-        }
-
-        
-    }
-}
