+++ Gestione/Models/interfaces.cs
@@ -15,8 +15,7 @@ namespace Interfaces{
         void AddCvStudi(string MatrCv,PerStud studi);
         void AddEspLav(string MatrCv, EspLav esp);
         void AddCompetenze(string MatrCv, Competenza comp);
-        void ModEspLav(string MatrCv, EspLav espV, EspLav esp );
-	
+        void ModEspLav(string MatrCv, EspLav espV, EspLav esp );	
 		void ModComp(Competenza daMod , Competenza Mod , string matricola); // Modifica la singola competenza
         void ModPerStudi(string matricola, PerStud daMod, PerStud Mod);
     }
